requirements.txt
Django==1.4
South==0.7.5
django-tagging==0.3.1
ipython==0.13.1
python-dateutil==2.1
rpy2==2.3.4
simplejson==3.1.2
six==1.3.0
wsgiref==0.1.2
